[REFACTOR] TagSelector component
